+++ ui/base/ime/ime_text_span.cc
@@ -11,8 +11,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace ui {
 
 ImeTextSpan::ImeTextSpan(Type type,
-                         uint32_t start_offset,
-                         uint32_t end_offset,
+                         size_t start_offset,
+                         size_t end_offset,
                          Thickness thickness,
                          UnderlineStyle underline_style,
                          SkColor background_color,
